refactor: Money 도메인 패키지 단순화
