+++ src/main/java/lotto/domain/money/Money.java
@@ -1,8 +1,5 @@
 package lotto.domain.money;
 
-import lotto.domain.money.condition.MoneyUnit;
-import lotto.domain.money.validation.MoneyValidator;
-
 public class Money {
 
     private final int money;
